WIP: debugging, more work on PoS, validation and keystore/provider

+++ src/test/skiplist_tests.cpp
@@ -77,7 +77,7 @@ BOOST_AUTO_TEST_CASE(getlocator_test)
     // Test 100 random starting points for locators.
     for (int n=0; n<100; n++) {
         int r = InsecureRandRange(150000);
-        CBlockIndex* tip = (r < 100000) ? &vBlocksMain[r] : &vBlocksSide[r - 100000];
+        CBlockIndex* tip = (r < 100000) ? &vBlocksMain[r] : &vBlocksSide[r - (uint64_t)100000];
         CBlockLocator locator = chain.GetLocator(tip);
 
         // The first result must be the block itself, the last one must be genesis.
